ThemeDecorator refactor CI 4
