+++ src/shared/ui/Loader/Loader.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { ThemeDecorator } from 'shared/config/storybookDecorators/ThemeDecorator'
 import { Theme } from 'app/providers/ThemeProvider'
+import { ThemeDecorator } from 'shared/lib/storybookDecorators/ThemeDecorator'
 import { Loader } from './Loader'
 
 const meta: Meta<typeof Loader> = {
