+++ src/widgets/LangSwitcher/ui/LangSwitcher.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { ThemeDecorator } from 'shared/config/storybookDecorators/ThemeDecorator'
 import { Theme } from 'app/providers/ThemeProvider'
+import { ThemeDecorator } from 'shared/lib/storybookDecorators/ThemeDecorator'
 import { LangSwitcher } from './LangSwitcher'
 
 const meta: Meta<typeof LangSwitcher> = {
