+++ src/widgets/PageError/ui/PageError.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { ThemeDecorator } from 'shared/config/storybookDecorators/ThemeDecorator'
 import { Theme } from 'app/providers/ThemeProvider'
+import { ThemeDecorator } from 'shared/lib/storybookDecorators/ThemeDecorator'
 import { PageError } from './PageError'
 
 const meta: Meta<typeof PageError> = {
